document-rag-chat/pyproject.toml
[tool.poetry]
name = "document-rag-chat"
version = "0.1.0"
description = "A Streamlit-based RAG application for chatting with documents"
authors = ["Ronen Artzi <ronen.artzi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
streamlit = "^1.31.0"
langchain = "^0.1.0"
langchain-community = "^0.0.16"
langchain-openai = "^0.0.5"
chromadb = "^0.4.22"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
ollama = "^0.3.3"
pypdf2 = "^3.0.1"
python-pptx = "^1.0.2"
python-docx = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"